MB-23573: Add full eviction defragment benchmarks
Enables the defragmenter benchmarks for full eviction. This required
changing the populate phase to use processSet rather than processAdd as
processAdd would check respond with bgFetch needed.

As a result of changing to processSet, when testing mapped memory we
need to clear the checkpointManager so that the memory gets released. I
also found that I was able to reduce the fuzz factor from 1.3 to 1.05.
Change-Id: I1cb01d9bb4501201caedd67c7da2e2fc1353b21a
Reviewed-on: http://review.couchbase.org/75971
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>