Provide a hardcoded get_log_info for mock engines
The DCP test suite in ep-engine don't use a mock_cookie as the
cookie causing the get_log_info to try to access bogus memory.

Until we've refactored the DCP tests to use a mock cookie
(it could stash the "this" pointer in the engine-specific
pointer) we can just return a hardcoded string.

Change-Id: I51bf0969fc93d400b8f1edcd6ff1e5c23b68f21b
Reviewed-on: http://review.couchbase.org/76030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>