Tests: Refactor to move helpers from MockEPVBucket to VBucketTest
Previously we accessed protected members of VBucket by inheriting from
it (MockEPVBucket), and adding public methods to the mock which called
the protected ones. This was fine when there was just one type of
VBucket to test, but this doesn't work well when we also want to test
other VBucket subclasses - e.g. EphemeralVBucket.

Similar to the changes made in 915ba08, change to making the test
fixture (VBucketTest) a friend of VBucket. This will allow the
soon-to-be-added EphemeralVBucketTest to inherit from VBucketTest, and
hence make use of all the same helper methods.

This also refactors the DefragmenterTest fixture so it inherits from
VBucketTest, as DefragmenterTest also used to use MockEPBucket and
hence needs updating to be able to setup its vbucket for testing.

Change-Id: I24da4098ea9da212284cd44a683a5e0e2e397fff
Reviewed-on: http://review.couchbase.org/75871
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>