MB-23637: findByName should not find connections marked as dead
When we attempt to make a connection with the same name we mark the
first as needing to disconnect and create a new connection with the
same name.  When manageConnections is executed the connection marked as
needing to disconnect will be closed.

However if findByName is called before manageConnections is run then it
could return either the new connection or the connection about to be
closed.

This patch changes findByName to only return the connection which is
not marked as needing to disconnect.

Change-Id: I02b77f30b3d48ca13e37fd5f07580c8fa3ea17e9
Reviewed-on: http://review.couchbase.org/76121
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>