MB-23454: Fix data race in createTapQueue
Appears to be caused by the same buggy string COW as MB-19688 and
MB-18510.

Upon updating the name of a tap connection, the old value was deleted.
This was considered by ThreadSanitizer to be racy with changes made in
one of multiple tasks.

Resolved in the same manner as MB-19688 - forcing an explicit
copy of `std::string name` in `ConnHandler::setName()`.

Change-Id: Iffee9ce2a695e97248e6b4f8aa640f4b1f4f3dfa
Reviewed-on: http://review.couchbase.org/76127
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>