[Ephemeral]: Serialize CMD_SEQNO_PERSISTENCE with vBucket seqno generation
This commit serializes the CMD_SEQNO_PERSISTENCE request w.r.t seqno
generation in ephemeral vbuckets because we must not queue up stale
requests. In regular couchbase vbuckets this is not a problem because
of cold runs of flusher task periodically. But in ephemeral we might end
up in a scenario wherein stale requests may never get notified as
there is no new sequence number triggering the notification.

It is not easy to write a test case for this race condition as we will
need 2 racing threads (writer thread and CMD_SEQNO_PERSISTENCE thread)
to be scheduled in a particular way inside the ep-engine.

Change-Id: Ie7ef1c1312b2047e834922f7c1dcb76491bd1588
Reviewed-on: http://review.couchbase.org/76195
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>