MB-23522: Refactor subdoc command flags for tests
This patch splits the flags in to path flags and doc flags and makes the
relevant changes at the call sites.

At the moment the two flag parameters are just logically or-ed together,
however this will change with a future patch which will seperate the
document flags in to a different type.

Change-Id: Iff6f2bc916b641bb60b0d3c338217e35561e74eb
Reviewed-on: http://review.couchbase.org/76245
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>