MB-23522: Add doc flags param to subdoc functions
Adds the ability to set doc flags seperately from path flags.

Changes the MKDOC and ACCESS_DELETED flags to be passed via the docFlags
parameter in the call sites.

Enables input validation for the addPathFlags method in
BinprotSubdocCommand so we can be sure we aren't passing doc flags as
path flags.

Change-Id: Ib68661806d1ab032415dc97b59e7515017e01fd5
Reviewed-on: http://review.couchbase.org/76249
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>