MB-23522: Split subdoc flags to path & doc flags
Introduces an enum class for the subdoc doc flags and removes them from
the subdoc flags. Most of the commit is the work associated with that as
well as few helper functions for the new enum class.

Introduces decoders to decode the optional doc flags from packets. This
involved updating the expected extras length for single and multi path
commands.

In the case where Mkdoc implies MKDIR_P, rather than pass another
argument (doc flags) down to the relevant functions I've just amended
the path flags such that MKDIR_P is set.

With regards to the tests, most of the changes and abstractions required
to deal with theconcept of doc flags have been made in previous commits.
This means that we really only change the actual types and values in
this commit.

Change-Id: I4d78d7d3d072fad3df883199671191ce2c62f190
Reviewed-on: http://review.couchbase.org/76403
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>