StoredValue: record deleted time for OSVs
For OrderedStoredValues, record the time when they are deleted, so
deleted OSVs can later be purged (completely deleted) when they reach
a certain age.

To avoid adding another 4bytes to OSV, the lock_expiry member is
re-used to also be used for the deleted time (deleted items cannot be
locked and hence this is otherwise unused once items enter that
state).

Change-Id: I1a6b9c5e8be4d3df1dca5c1e91be6219889686da
Reviewed-on: http://review.couchbase.org/76406
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>