[Ephemeral]: Rollback to 0 explicitly and improve rollback testing
In ephemeral buckets we currently rollback to 0 even if the request
is to rollback to an intermediate point. This is because we don't have
the luxury of disk snapshots to do an intermediate rollback.

This commit does an explicit rollback to 0, rather than as default.

Also adds a test case that requests a rollback to an intermediate
point.

Change-Id: Ifea06562efaae4ef9b995c718a1b1adc026c4e02
Reviewed-on: http://review.couchbase.org/76444
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>