MB-23836: Use LIBEVENT_LIBRARIES instead of LIBEVENT_CORE_LIB
If the user is trying to build on a non-supported platform and
is trying to use libevent built with autoconf we must link with
libevent_pthreads in addition to the other libraries.

To make our life easier we'll hide all of that magic in the
cmake file in tlm which search for all of the libraries,
and link with all of the libevent libraries instead.

Change-Id: I991bc11bdc5fc8e66bf8afa87e5af569d20cd9b5
Reviewed-on: http://review.couchbase.org/76682
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>