MB-23522: Allow Add semantics for multi mutations
This patch adds a new doc flag "Add" which, when set, provides Add
semantics for subdoc mutations. These are: fail if the key exists or
if the cas of the request is set.

The addition of this flag means that it is now possible for subdoc
mutations to exhibit 3 mutation semantics:
  Replace - Where neither Mkdoc or Add flags are set
  Set - Where the Mkdoc flag is set
  Add - Where the Add flag is set

Using the Add and Mkdoc flags together is currently invalid, however
this can be changed in the future should we desire some fourth mutation
semantic.

Change-Id: I6493212213020a2ec5240206171e18fca83b93fd
Reviewed-on: http://review.couchbase.org/76707
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>