MB-23875: Implement GAT and TOUCH in the core [2/2]
The motivation for this is to decouple the protocol encoding
of the commands from the engine implementation, and to fix
problems with these commands not being included in the
global counters for the command count.

As an extra benefit we can add end to end test through
commit validation as it is supported in all engines.

Change-Id: I2a26987c16b186c91de6f2bc7ab2b87d9c53ec67
Reviewed-on: http://review.couchbase.org/76936
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>