MB-23377: Move GetValue ownership for multi fetch
GetValues currently have a raw pointer to the blob which they are
associated with. Really they should own this and pass on ownership when
required to ensure that resources are properly cleaned up.

However in the case of multi bg fetches multiple VBucketBGFetchItems may
have GetValues which refer the same underlying blob. Therefore in this
case there is an n X 1 relationship between VBucketBGFetchItems and
underlying blobs (the things we are fetching). Therefore it would be
incorrect for a VBucketBGFetchItem to own something which owns one of
these blobs.

This patch moves the ownership of GetValues from VBucketBGFetchItems to
vb_fetch_item_ctx_ts which typically has a  vector of
VBucketBGFetchItems. The VBucketBGFetchItems then hold a pointer to the
GetValue.

Change-Id: I48b2dc8535f0f7eb2beb9eaa9ed504d1b5a2b7ca
Reviewed-on: http://review.couchbase.org/78598
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>