MB-23909: Add error object to error response messages
The payload of an error message is changed to:

   { "error":
     {
       "ref":"00000000-0000-0000-0000-000000000000",
       "context":"All work and no play makes Jack a dull boy"
     }
   }

And the JSON bit is set in the datatype field. Both fields
in the error object is optional.

ref is only present if it is written to the log
context is only present if it contains more context information
than the status code (We'll not transfer "Not found", as that
offers no benefit to the caller, but it may contain information
such as:

"Authorization failure: can’t execute SET_WITH_META
 operation without the MetaWrite privilege."

Change-Id: I121e57d898a8f0d50a356cde9cb2fa8e25f47a6b
Reviewed-on: http://review.couchbase.org/77225
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>